Move vision encoders and fusion modules into separate files
