+++ aegis_gym/rsl/behavior_cloning.py
@@ -3,7 +3,7 @@ import time
 from collections import deque
 from collections.abc import Iterator
 from pathlib import Path
-from typing import Callable, Optional
+from typing import Optional
 
 import numpy as np
 import torch as th
@@ -12,6 +12,19 @@ import torch.nn.functional as F
 import torchvision.utils as vutils
 from clearml import Task
 from rsl_rl.utils.logger import Logger
+
+from bc_encoders import (
+    VisionEncoder,
+    SharedCNNEncoder,
+    PerCameraCNNEncoder,
+    AutoencoderCNNEncoder,
+)
+from bc_fusions import (
+    FusionModule,
+    LinearFusion,
+    VectorAttentionFusion,
+    SpatialAttentionFusion,
+)
 
 
 class BehaviorCloning:
@@ -639,310 +652,3 @@ class Policy(nn.Module):
             input_dim = h
         layers.append(nn.Linear(input_dim, output_dim))
         return nn.Sequential(*layers)
-
-
-class VisionEncoder(nn.Module):
-    def __init__(self, num_cameras: int):
-        super().__init__()
-        self.num_cameras = num_cameras
-
-    def forward(self, rgb_obs: th.Tensor) -> tuple[th.Tensor, ...]:
-        raise NotImplementedError
-
-    def infer_output_shape(
-        self,
-        image_height: int = 64,
-        image_width: int = 64,
-        device: str = "cpu",
-    ) -> tuple[int, int, int]:
-        dummy = th.zeros((1, 3, image_height, image_width), device=device)
-        with th.no_grad():
-            out = self._single_forward(dummy)  # (1, C, H, W)
-        _, c, h, w = out.shape
-        return c, h, w
-
-    def _single_forward(self, x: th.Tensor) -> th.Tensor:
-        raise NotImplementedError
-
-
-class FusionModule(nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    def forward(self, features: tuple) -> th.Tensor:
-        raise NotImplementedError
-
-    def prepare_pose_features(self, features: tuple) -> tuple:
-        raise NotImplementedError
-
-
-class SharedCNNEncoder(VisionEncoder):
-    def __init__(self, num_cameras: int, cnn_builder: Callable, vision_cfg: dict):
-        super().__init__(num_cameras)
-        self.encoder = cnn_builder()
-
-    def _single_forward(self, x: th.Tensor) -> th.Tensor:
-        return self.encoder(x)
-
-    # TODO(issue#71): Investigate indexing and micro-optimizations in vision encoder forward pass
-    def forward(self, rgb_obs: th.Tensor) -> tuple[th.Tensor, ...]:
-        features = []
-        for i in range(self.num_cameras):
-            cam_rgb = rgb_obs[:, i * 3 : (i + 1) * 3]
-            feat = self.encoder(cam_rgb)
-            features.append(feat)
-        return tuple(features)
-
-
-class PerCameraCNNEncoder(VisionEncoder):
-    def __init__(self, num_cameras: int, cnn_builder: Callable, vision_cfg: dict):
-        super().__init__(num_cameras)
-        self.encoders = nn.ModuleList([cnn_builder() for _ in range(num_cameras)])
-
-    def _single_forward(self, x: th.Tensor) -> th.Tensor:
-        return self.encoders[0](x)
-
-    # TODO(issue#71): Investigate indexing and micro-optimizations in vision encoder forward pass
-    def forward(self, rgb_obs: th.Tensor) -> tuple[th.Tensor, ...]:
-        features = []
-        for i in range(self.num_cameras):
-            cam_rgb = rgb_obs[:, i * 3 : (i + 1) * 3]
-            feat = self.encoders[i](cam_rgb)
-            features.append(feat)
-        return tuple(features)
-
-
-# TODO(issue#79): Remove hardcoded dimensions related to autoencoder and make them configurable
-class AutoencoderCNNEncoder(VisionEncoder):
-    def __init__(self, num_cameras: int, cnn_builder: Callable, vision_cfg: dict):
-        super().__init__(num_cameras)
-
-        self.encoder = cnn_builder()
-
-        self.feature_channels = vision_cfg["conv_layers"][-1]["out_channels"]
-        self.feature_size = 16
-        self.flatten_dim = self.feature_channels * self.feature_size * self.feature_size
-        self.latent_dim = 512
-
-        self.to_latent = nn.Linear(self.flatten_dim, self.latent_dim)
-        self.from_latent = nn.Linear(self.latent_dim, self.flatten_dim)
-        self.feature_channels = self.feature_channels
-        self.feature_size = self.feature_size
-        self.decoder = self._build_decoder()
-
-    def _single_forward(self, x: th.Tensor) -> th.Tensor:
-        fmap = self.encoder(x)  # (1, 32, 16, 16)
-        latent = self.to_latent(fmap.flatten(start_dim=1))  # (1, 512)
-        return latent.unsqueeze(-1).unsqueeze(-1)  # (1, 512, 1, 1)
-
-    def forward(self, rgb_obs: th.Tensor) -> tuple[th.Tensor, ...]:
-        features = []
-        for i in range(self.num_cameras):
-            cam_rgb = rgb_obs[:, i * 3 : (i + 1) * 3]
-            fmap = self.encoder(cam_rgb)  # (B, 32, 16, 16)
-            latent = self.to_latent(fmap.flatten(start_dim=1))  # (B, 512)
-            features.append(latent.unsqueeze(-1).unsqueeze(-1))  # (B, 512, 1, 1)
-        return tuple(features)
-
-    def reconstruct(self, rgb_obs: th.Tensor) -> tuple[th.Tensor, ...]:
-        recons = []
-        for i in range(self.num_cameras):
-            cam_rgb = rgb_obs[:, i * 3 : (i + 1) * 3]
-            fmap = self.encoder(cam_rgb)  # (B, 32, 16, 16)
-            flat = fmap.flatten(start_dim=1)
-            latent = self.to_latent(flat)  # (B, 512)
-            flat_recon = self.from_latent(latent)  # (B, 8192)
-            fmap_recon = flat_recon.view(
-                -1, self.feature_channels, self.feature_size, self.feature_size
-            )  # (B, 32, 16, 16)
-            recons.append(self.decoder(fmap_recon))  # (B, 3, 64, 64)
-        return tuple(recons)
-
-    def _build_decoder(self) -> nn.Sequential:
-        return nn.Sequential(
-            # (B, 32, 16, 16)
-            nn.ConvTranspose2d(
-                32, 16, kernel_size=3, stride=2, padding=1, output_padding=1
-            ),
-            nn.ReLU(),
-            # (B, 16, 32, 32)
-            nn.ConvTranspose2d(
-                16, 8, kernel_size=3, stride=2, padding=1, output_padding=1
-            ),
-            nn.ReLU(),
-            # (B, 8, 64, 64)
-            nn.Conv2d(8, 3, kernel_size=3, stride=1, padding=1),
-            nn.Sigmoid(),
-            # (B, 3, 64, 64)
-        )
-
-
-class LinearFusion(FusionModule):
-    def __init__(
-        self,
-        vision_dim: int = 512,
-        num_cameras: int = 3,
-        in_channels: int = 32,
-        image_height: int = 64,
-        image_width: int = 64,
-        pool_size: int = 4,
-    ):
-        super().__init__()
-        self.pool_size = pool_size
-        self.is_latent_vector = image_height == 1 and image_width == 1
-
-        if self.is_latent_vector:
-            feature_dim_cam = in_channels
-            self.pose_proj = nn.Linear(in_channels, vision_dim)
-        else:
-            feature_dim_cam = in_channels * pool_size * pool_size
-            self.pose_proj = None
-
-        self.net = nn.Sequential(
-            nn.Linear(feature_dim_cam * num_cameras, vision_dim),
-            nn.ReLU(),
-            nn.Dropout(0.1),
-        )
-        self.output_dim = vision_dim
-        self.pose_input_dim = feature_dim_cam
-
-    def _to_feature_vector(self, feat: th.Tensor) -> th.Tensor:
-        if self.is_latent_vector:
-            f = feat.flatten(start_dim=1)
-        else:
-            f = F.adaptive_avg_pool2d(feat, (self.pool_size, self.pool_size))
-            f = f.flatten(start_dim=1)
-        return f
-
-    def forward(self, features: tuple) -> th.Tensor:
-        feat_vecs = [self._to_feature_vector(f) for f in features]
-        fused_feat_vec = th.cat(feat_vecs, dim=-1)
-        return self.net(fused_feat_vec)
-
-    def prepare_pose_features(self, features: tuple) -> tuple:
-        feat_vecs = [self._to_feature_vector(f) for f in features]
-        if self.pose_proj is not None:
-            return tuple(self.pose_proj(f) for f in feat_vecs)
-        else:
-            return tuple(feat_vecs)
-
-
-class VectorAttentionFusion(FusionModule):
-    def __init__(
-        self,
-        vision_dim: int = 512,
-        num_cameras: int = 3,
-        in_channels: int = 32,
-        num_heads: int = 4,
-        pool_size: int = 4,
-    ):
-        super().__init__()
-        self.pool_size = pool_size
-        flat_per_cam = in_channels * pool_size * pool_size
-
-        self.proj_in = nn.Linear(flat_per_cam, vision_dim)
-        self.cam_embed = nn.Parameter(th.randn(1, num_cameras, vision_dim) * 0.02)
-        self.cls_token = nn.Parameter(th.randn(1, 1, vision_dim) * 0.02)
-        self.norm = nn.LayerNorm(vision_dim)
-        self.attn = nn.MultiheadAttention(
-            vision_dim, num_heads, batch_first=True, dropout=0.0
-        )
-        self.mlp = nn.Sequential(nn.Linear(vision_dim, vision_dim), nn.GELU())
-
-        self.output_dim = vision_dim
-        self.pose_input_dim = flat_per_cam
-
-    def _to_feature_vector(self, feat: th.Tensor) -> th.Tensor:
-        f = F.adaptive_avg_pool2d(feat, (self.pool_size, self.pool_size))
-        return f.flatten(start_dim=1)
-
-    def forward(self, features: tuple) -> th.Tensor:
-        feat_vecs = [self.proj_in(self._to_feature_vector(f)) for f in features]
-
-        x = th.stack(feat_vecs, dim=1)
-        x = x + self.cam_embed
-
-        B = x.shape[0]
-        cls = self.cls_token.expand(B, -1, -1)
-        x = th.cat([cls, x], dim=1)
-
-        x_norm = self.norm(x)
-        attn_out, _ = self.attn(x_norm, x_norm, x_norm)
-        x = x + attn_out
-
-        cls_out = x[:, 0]
-        return self.mlp(cls_out)
-
-    def prepare_pose_features(self, features: tuple) -> tuple:
-        return tuple(self._to_feature_vector(f) for f in features)
-
-
-class SpatialAttentionFusion(FusionModule):
-    def __init__(
-        self,
-        vision_dim: int = 256,
-        num_cameras: int = 3,
-        in_channels: int = 64,
-        image_height: int = 8,
-        image_width: int = 8,
-        num_heads: int = 4,
-    ):
-        super().__init__()
-        num_patches = image_height * image_width
-
-        self.patch_proj = nn.Linear(in_channels, vision_dim)
-        self.pos_embed = nn.Parameter(th.randn(1, num_patches, vision_dim) * 0.02)
-        self.cam_embed = nn.Parameter(th.randn(1, num_cameras, 1, vision_dim) * 0.02)
-        self.cls_token = nn.Parameter(th.randn(1, 1, vision_dim) * 0.02)
-
-        self.norm1 = nn.LayerNorm(vision_dim)
-        self.attn = nn.MultiheadAttention(
-            vision_dim, num_heads, batch_first=True, dropout=0.0
-        )
-        self.norm2 = nn.LayerNorm(vision_dim)
-        self.mlp = nn.Sequential(nn.Linear(vision_dim, vision_dim), nn.GELU())
-
-        self.spatial_pool = SpatialAttentionPooling(in_channels)
-
-        self.output_dim = vision_dim
-        self.pose_input_dim = in_channels
-
-    def forward(self, features: tuple) -> th.Tensor:
-        B = features[0].shape[0]
-        patch_tokens_per_cam = []
-
-        for i, feat in enumerate(features):
-            patches = feat.flatten(2).transpose(1, 2)
-            patch_tokens = self.patch_proj(patches)
-            patch_tokens = patch_tokens + self.pos_embed
-            patch_tokens = patch_tokens + self.cam_embed[:, i]
-            patch_tokens_per_cam.append(patch_tokens)
-
-        x = th.cat(patch_tokens_per_cam, dim=1)
-
-        cls = self.cls_token.expand(B, -1, -1)
-        x = th.cat([cls, x], dim=1)
-
-        x_norm = self.norm1(x)
-        attn_out, _ = self.attn(x_norm, x_norm, x_norm)
-        x = x + attn_out
-
-        cls_out = x[:, 0]
-        return self.mlp(self.norm2(cls_out))
-
-    def prepare_pose_features(self, features: tuple) -> tuple:
-        return tuple(self.spatial_pool(f) for f in features)
-
-
-class SpatialAttentionPooling(nn.Module):
-    def __init__(self, in_channels: int):
-        super().__init__()
-        self.conv = nn.Conv2d(in_channels, 1, kernel_size=1)
-
-    def forward(self, x: th.Tensor) -> th.Tensor:
-        B, C, H, W = x.shape
-        weights = self.conv(x)
-        weights = weights.view(B, 1, H * W)
-        weights = th.softmax(weights, dim=-1)
-        feats = x.view(B, C, H * W)
-        return (feats * weights).sum(dim=-1)
